requirements/local.txt
# Local development dependencies go here
-r base.txt

coverage==4.5.3
django-coverage-plugin==1.6.0

Sphinx==2.0.0
django-extensions==2.1.6
Werkzeug==0.15.1
django-test-plus==1.1.1
factory-boy==2.11.1

django-debug-toolbar==1.11

# improved REPL
ipdb==0.12

pytest-django==3.4.8
pytest-sugar==0.9.2
